requirements.txt
python-dotenv
pydash
discord
discord.py[voice]
elasticsearch
openai
humanize
google-cloud-texttospeech
